Extract creation of the picker to a separate class, since there is a delegate that has to be implemented
